+++ src/v2/providers/alerts/crashlytics.ts
@@ -245,7 +245,7 @@ export interface CrashlyticsOptions extends options.EventHandlerOptions {
   /**
    * Connect cloud function to specified VPC connector.
    */
-  vpcConnector?: string | ResetValue;
+  vpcConnector?: string | Expression<string> | ResetValue;
 
   /**
    * Egress settings for VPC connector.