+++ src/v2/providers/eventarc.ts
@@ -130,7 +130,7 @@ export interface EventarcTriggerOptions extends options.EventHandlerOptions {
   /**
    * Connect cloud function to specified VPC connector.
    */
-  vpcConnector?: string | ResetValue;
+  vpcConnector?: string | Expression<string> | ResetValue;
 
   /**
    * Egress settings for VPC connector.