+++ CHANGELOG.md
@@ -1,1 +1,3 @@
+- Allow Functions to specify vpcConnector as a string parameter (#1329)
+- Upgrade jsonwebtoken to version 9 (#1336)
 - Adds 'eventarcpublishing' as required API to custom event function